Update MainActivity.kt
Handle Advertising
In the BluetoothManager, the startAdvertising() method initiates BLE advertising, broadcasting a custom UUID. This can be changed to broadcast any type of data you need to share with nearby devices. The AdvertiseCallback lets you know if the advertising started successfully or failed.

Success: Displays a toast indicating advertising has started.
Failure: Displays a toast with an error code if the advertising fails.