Update BluetoothManager.kt
Sending Signal: In the sendSignal method, We write the signal (e.g., "CHANGE_COLOR") to the connected device's GATT characteristic.
Receiving Signal: The onCharacteristicChanged method listens for changes in the characteristic, which allows the device to receive the signal. When a signal is received, wecall handleReceivedSignal to perform the corresponding action (like changing color or playing a sound).