Update MainActivity.kt
handleReceivedSignal(signal: String): This function handles the received signal. Depending on the signal type (CHANGE_COLOR or SOUND_ALERT), it will either change the background color of the screen or play the siren sound.

playSirenSound(): This function initializes a MediaPlayer object to play the siren.mp3 file. The sound is played when the SOUND_ALERT signal is received.